requirements.txt
selenium
requests
boto3
python-dotenv
